Not nearly done, just hacking.
