+++ src/westmere/stringparsing.h
@@ -7,6 +7,7 @@
 #include "simdjson/parsedjson.h"
 #include "jsoncharutils.h"
 #include "westmere/intrinsics.h"
+#include "westmere/bitmanipulation.h"
 
 #ifdef JSON_TEST_STRINGS
 void found_string(const uint8_t *buf, const uint8_t *parsed_begin,
